examples/nextjs: server components + Suspense streaming for home & header
Restructure the Next.js example toward better App Router patterns (and
toward Cache Components):

- Extract CartButton as a client island so Header can be a server
  component that fetches its own collections.
- Split the home product grid into a static shell + a Suspense-streamed
  ProductList with a skeleton fallback.
- Drop force-dynamic and the blocking cart/collections fetch from the
  root layout; data is colocated in the components that need it and the
  request-scoped storefront client stays React.cache()-memoized.
- Cart is seeded client-side by CartProvider's built-in fetch instead of
  a blocking server seed, keeping the layout non-blocking. See the
  Providers TODO for the promise + React.use() path that needs core
  support.
- Load standard-actions.js with afterInteractive so next/script injects
  it via an effect rather than rendering a <script> React 19 won't
  execute on the client.
