+++ examples/nextjs/app/layout.tsx
@@ -1,4 +1,3 @@
-import { HEADER_COLLECTIONS_QUERY, normalizeHeaderCollections } from "@shared/header";
 import type { Metadata } from "next";
 
 import "./globals.css";
@@ -11,8 +10,6 @@ import { CartDrawer } from "@/components/CartDrawer";
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
 import { Providers } from "@/components/Providers";
-import { cartHandlers } from "@/lib/cart-handlers";
-import { getStorefrontClient } from "@/lib/storefront";
 
 const inter = Inter({
   subsets: ["latin"],
@@ -24,29 +21,12 @@ export const metadata: Metadata = {
   title: "Mock.shop — Hydrogen",
 };
 
-export const dynamic = "force-dynamic";
-
-export default async function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
-  const storefrontClient = await getStorefrontClient();
-  const [{ data: cartData }, { data: headerData }] = await Promise.all([
-    cartHandlers.get({ storefrontClient }),
-    storefrontClient.graphql(HEADER_COLLECTIONS_QUERY),
-  ]);
-  const headerCollections = normalizeHeaderCollections(headerData?.collections?.nodes);
-
+export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
     <html lang="en" className={inter.variable}>
-      <head>
-        <Script
-          src="https://cdn.shopify.com/storefront/standard-actions.js"
-          type="module"
-          crossOrigin="anonymous"
-          strategy="beforeInteractive"
-        />
-      </head>
       <body className="bg-white text-black">
-        <Providers cart={cartData.cart}>
-          <Header collections={headerCollections} />
+        <Providers>
+          <Header />
           <Suspense fallback={null}>
             <AnalyticsTracker />
           </Suspense>
@@ -54,6 +34,12 @@ export default async function RootLayout({ children }: Readonly<{ children: Reac
           <Footer />
           <CartDrawer />
         </Providers>
+        <Script
+          src="https://cdn.shopify.com/storefront/standard-actions.js"
+          type="module"
+          crossOrigin="anonymous"
+          strategy="afterInteractive"
+        />
       </body>
     </html>
   );
